change prices from <p> to span and reposition
I wanted the prices to be on the same line as the menu item and lined
up vertically. To achieve this, I moved the prices from their own
paragraph to a span class on the same line as the <h3> menu item, then
I positioned them absolutely on the page at 50% to the left.