add 'half' and 'nothalf' classes to menu sections
I added images to the sides of the menu sections. At min-width:400px,
some awkward white spaces are created by the "Burgers and "Sides"
sections. To alleviate this, I created a class of "half" to use for
those sections with images that sets the section at a width of 50% when
the viewport is at min-width:400. Then, the sections that are "nothalf"
can flow the entire width of the container.

Actually, I'll bet that I don't need the 'nothalf' class at all.