remove superfluous 'nothalf' class
As I suspected, I didn't need the 'nothalf' class to ensure that the
text would flow through the entire width of the container. I don't know
what I was thinking. Guess I'm getting tired.

Oh, on that last commit, I also changed a span class from 'half' to
'halfoff' to avoid confusion. Again, the main point is that I didn't
want any weird white space as a result of the way the images are
positioned when the viewpoint is at min-width:400px.