move "About Us" link to top and remove zucchini img
I moved the About Us link to the top of the hyperlinks list to match
the order of the sections and removed the zucchini image and all
related CSS styling. Also removed CSS styling for the Chive img I
previously removed and some superfluous #BOTD styling that I didn't
need as it was taken care of by other styling.